Migrate all try/except blocks to the syntax specified in PEP 3110[0] and supported in Python 2.6 and later.
TL;DR: If you use AppEngine's 2.5 runtime or 2.4 from GRTE, don't
approve this, ask me to revert and read the rest of this description.

The old syntax looks like this:

try:
...
except (ErrorType1, ErrorType2), e:
...

while the new format looks like this:

try:
...
except (ErrorType1, ErrorType2) as e:
...

This reduces the likelihood of wanting to catch two exception types
but accidentally catching only one of the types and using the second
exception type as the variable name. It's now recommended by the style
guide [1] and is forward-compatible with Python 3.x. This will ease an
eventual transition to Python 3, which is currently expected to
begin in 2013.

NOTE: This CL is being managed by http://go/rosie, and is part of a
batch that touches just over 18,000 files. If you have nits that are
unrelated to this change, I'd _greatly_ appreciate it if you could fix
them yourself in a followup. If your project needs to retain 2.4
compatibility, please ask me to revert these changes, but know that
GRTE versions which contain Python 2.4 are very close to being
completely unsupported, and python-team really wants people on 2.6
ASAP. Python 2.7 will be part of the upcoming GRTEv3.

[0]: http://www.python.org/dev/peps/pep-3110/
[1]: http://www.corp.google.com/eng/doc/pyguide.xml#Exceptions

Tested:
TAP presubmits for global presubmit queue

R=csharp,mad
DELTA=23  (0 added, 0 deleted, 23 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=5062


git-svn-id: https://swarming.googlecode.com/svn/trunk@106 722f4569-f0a1-f026-5e06-c3648ed4afc7