Enable canonical representation of IP addresses.
Make WhitelistedIP use the canonical IP representation as the key so save a
member lookup. This will be useful when support for subnet is added.

Make it possible to add an IP address that is not the client ip.

Add first ever unit test.

R=csharp@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/14333017

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/isolate_server@196224 0039d316-1c4b-4281-b951-d872f2087c98