Add HTTP-level caching support. It significantly help with download performance.
Add -deflate as an alias for -gzip. Add TODOs to get rid of -gzip since it is
ill-named. Add a note that it's impossible to specify Content-Encoding within an
AppEngine instance because it's bone-head. That would enable XHR requests to
have the browser decompress the file on the fly, but that's just too bad.

R=iannucci@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/17449005

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/isolate_server@207432 0039d316-1c4b-4281-b951-d872f2087c98