Use NDB's get_async in /content-gs/pre-upload call instead of DB query.
It pulls in whole entity in the memory but reuses NDB memcache (instead of
hitting the datastore all the time). Net gain is ~40% faster check for
a warm cache. We might want to implement our own caching scheme for
ContentEntry presence bit in the future if memory/performance of NDB
based approach becomes an issue.

BUG=
R=maruel@chromium.org

Review URL: https://codereview.appspot.com/14414043