Move the handlers into handlers.py to make it possible to unit test it.
The bootstrap() function should not be called while importing the modules, so
they can be tested properly. So contain side effects when importing to main.py.
This way, the handlers can safely be unit tests.

It was a problem with stats, where datastore access was done during the
construction of _STATS_HANDLER.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/15060050