Correct problem where start_time or end_time is given 0.
Replace 0 with None, so that the proper default values are used.

Fix generate_report() to not alter arguments start_time and end_time, so they
are not different values in the generated template.

Remove start_time and end_time as arguments to generate_and_email_report(), they
are never used and not meant to be used.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/19140043