Make the html templates more HTML5 friendly.
Change app.yaml to use a wholesale route for static files instead of listing
each files one by one. Adds support for /robots.txt, it was missing. Stop using
images/ and css/ subdirectories. A directory for a single file is not useful.

Remove the use of 'cycle' in html templates. This keyword exists in django but
not in jinja2. This is replaced with a CSS selector.

Reformat the html templates to be mostly 80 cols and indent with regards to the
objects being created.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/34140043