Enable appstats on swarming on dev instances.
Rework the bootstrapping to use the environment variable APPLICATION_ID instead
of app_identity.get_application_id(), and d it later in the boot process.

Calling this function does an RPC request, which could hang during the
bootstrap. We are seeing hangs during startup so it could well be related.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/35960043