Change version calculation to ignore submodules when checking if it is tainted.
Otherwise if client is not exactly at the submodule version, '-tainted' will
always be appended to the version name, which is annoying and unnecessary. The
fact that client is at any version, or not even checked out, doesn't have any
effect on the server code.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/36360045