Stop modifying entities in an handler.
Doing something like "machine.status = 'foo'" is bad because two concurrent
handlers will get the same in-memory object instance for the same entity if it
is kept in the local cache, causing incoherency in the objects. So objects
must not be modifying 'in-place' for UI purposes. To guarantee that, the
entities are converted to dict, where it is modified for UI purposes. In short,
the code was not thread-safe.

Remove MachineStats.last_seen_str, presentation should not be in the model.
Create a jinja2 custom filter instead so UI formatting is in the HTML template.

Make SortOptions a namedtuple, it takes less code and is more memory efficient.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/36630043