Explicitly handle a few datastore_errors exceptions while tagging.
When the datastore feels like it's Friday and starts to misbehave, it generates
an incredible amount of spew. Reduce the noise about things we can't do anything
about and that won't break the immediate functionality. If the task queue fails,
it may cause the items to be evicted sooner but that is just it, and the task
queue will be retried automatically.

R=vadimsh@chromium.org
BUG=chromium:341734

Review URL: https://codereview.appspot.com/61630046