Add coverage tool.
Automates the annoying part of setupping GAE SDK path and then stripping it from
the report.

Rename post_test.py to post_task.py so it doesn't have to be blacklisted, it is
not a test.

Fix local_test_runner_test.py to be less noisy.

Fix swarm_generator.py file mode, it is not executable.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/66080043