Add skeleton of pkg/aedmz and isolateserver in Go.
This implements the barebone HTTP router so functionality can be incrementally
piled on.

aedmz is the abstraction layer to make it possible to run the same code
functional both on a standalone server an on AppEngine. It does make the code
more complicated but the advantage is significant for testing rapidly.

R=csharp@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/49110043