Remove output_destination and support for streaming to an arbitrary server.
This wasn't used but will be needed back. In the meantime remove it since it is
a liability. When streaming support is added back to TestRun, it will replace
result_url, not be _in addition to_ result_url. It's not efficient to have two
ways to upload logs back, so only keep the one known to work.

Enforce result_url and encoding in TestRun.

Use more sensible values in some tests, for example an environment variable
value must be a string, it can't be a int.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/70040044