Modify /get_slave_code to optionally accept slave code version.
This version is verified to be valid, so the right code is downloaded. This may
solve issues where bots are not running the expected swarm_bot code.

Use Content-Disposition, so that downloading with curl or a web browser will
result in a properly named file.

Add corresponding unit test.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/74040044