Remove both .store_result and .result_url from TestCase.
TestCase.store_result is now always 'all'. Results are always stored on the
Swarming master. Other values do not make sense.

TestCase.result_url was the remaining side effect in TestCase. Now a TestCase is
idempotent, rerunning the same TestCase won't have any side effect outside the
Swarming infrastructure.

Note that TestRun.result_url is not affected, it has a different purpose and is
kept.

Incidentally, this removes the last active usage of url_helper on the server.
The file will be moved in a following CL.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/70210045