Implement Logger and LogService in aedmz.
Logger permits HTTP request specific logging.
LogService permits reading the logs back.

Only test the local implementation, the one on GAE just works and the local stub
is not worth being tested.

Make aedmztest.NewAppMock() match aedmz.NewAppMock() so in the future it is
easier to fake time for example. On the other hand, rename aedmz.NewAppMock() to
newAppMock so users are not confused and may think it is exported.

R=csharp@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/70140044