Add components/sharding.py module to define a clear way to shard entities.
The module is strikingly simple, it defines a single function to get a key to a
shard. It can't get any simpler. Yet this is all is needed to shard efficiently,
since the entity itself doesn't need to exist in the db.

Remove ContentNamespace, it is not referenced anymore.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/77320044