Stop modifying entities in memory.
Both TestRunner and SwarmError were modified in memory. This can result in
problems when the objects are cached in process. It's better to not modify the
entity in views.

Extract GetTimeString() out of handler into template.py where it belongs.

Extract _GetDisplayableRunnerTemplate() out of the class into
make_runner_view(). It'd be better to put it into template but it'd be more
tricky to do so postpone.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/78800048