Ensure tests in auth_ui_rest_api_test.py perfom all expected permission checks.
This prevents situations when a call used to use some permission, but not using
it anymore (due to a legitimate change or a bug in code), but test still passes.

Also add a clarifying comment to @require decorator that explains how multiple
stacked @require decorators behave.

R=maruel@chromium.org
BUG=

Review URL: https://codereview.appspot.com/80240043