auth: REST API endpoint and UI for group deletion.
A group can be deleted only if nothing depends on it which means
it's not referenced in any ACL rules and not used as a nested group
in another group.

Also support 'If-Unmodified-Since' condition. It is used to guarantee
that no changes are overwritten when multiple users modify a group
concurrently via UI.

R=maruel@chromium.org
BUG=

Review URL: https://codereview.appspot.com/81230044