Rename task_glue.py _USE_OLD_API to USE_OLD_API.
It will have to be referenced by other files; for example some unit tests access
directly the DB so they have parts that need to be skipped out.

Note that the end goal is to completely delete task_glue.py but the abstraction
layer is useful during the transition so I can compare the behavior of the APIs
between the old and new ones.

R=phajdan.jr@chromium.org
BUG=swarming:89

Review URL: https://codereview.appspot.com/87770044