Add compression support while keeping backward compatibility.
I may start adding a lot of compressible junk in the snapshots, so it worth
compressing them as a big blob.

Backward compability is preferable to not break the data currently on the
isolate server. A mapping job could be done to convert the remaining data if I
wasn't lazy.

R=csharp@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/98760044