Move make_runner_view and GenerateButtonWithHiddenForm to task_glue.py.
GenerateButtonWithHiddenForm should be done in JS but punting this for later.
make_runner_view needs knowledge about the internal object so move it to
task_glue to key it on USE_OLD_API. This one will have to be reimplemented to
jinja2 based templating later.

R=csharp@chromium.org
BUG=swarming:89

Review URL: https://codereview.appspot.com/90830043