Move stats/machine_stats.py as server/bots_list.py.
I'm changing the word 'bot' than 'machine'. This CL is part of the process to
use consitent naming throughout the code base.

Technically, this is not statistics, it's really just a flat list, that's why
thr file is being renamed. I didn't change the content much so the history is
clean but the next CL will refactor the code and reenable its use.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/101790044