Remove appstats from isolate server as an experimentation.
It is possible it may be causing bursty startup failures we are observing from
times to time. This CL is to try it out on prod for at least several days to see
if it is the actual cause or at least contributes to make this problem more
significant.

The observed logged error is usually an ImportError exception accompanied with:
"Process terminated because the request deadline was exceeded during a loading
request."

TBR=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/103860043