Make swarming_bot.zip immediately self-replicate on first execution.
This makes it more resilient in case of partial destruction, which happened on
the bots. This is *not* strictly necessary but the zip is relatively small, in
the ~100kb range, so it is not too much of a big deal. Will revisit if it
becomes a liability.

R=csharp@chromium.org
BUG=swarming:109

Review URL: https://codereview.appspot.com/93680043