Try to improve Windows bots resilience and transfer restart control to __main__.
setup_bot() returning False will cause the caller, in CMDstart_slave, to restart
the host on behalf of start_slave.py. This simplifies the update flow, because
in this case it is not needed to restart the host.

Windows bots tend to fail to update in a weird state where conhost.exe is still
running but not cmd.exe. #thisiswindows. This CL tries to shed some light on the
failures:

- Change Windows automatic startup command to log to a text file.
- Better logging when updating to figure out windows specific failure.
- Use subprocess.Popen() instead of os.execv(), it'll likely be more
  Windows-friendly.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/103880044