Warmup the slave bot version and zip, save it in memcache.
When a new version is deployed, this value is calculated for each bot doing a
concurrent request so pre-calculating it and saving it in memcache is going to
help a bit.

The zip was never saved in memcache, when scaling to 1000 bots, it's worth
saving this duplicate calculation.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/103070044