Saves currently running task on bot list.
Stop spacing Bot update every 2 minutes, update when the bot pings, which is
less that once a second. Each bot is a root entity so in theory they should
sustain a 1 write/sec correctly, in practice it should be lower than that. If it
becomes a problem, internal will be added back.

Remove IPs as columns from the list. It'll be shown on a (soon to be added) per
bot page.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/108820043