Make the bot more resilient and add more Windows dimensions.
The Windows integrity level is to know if the process has administrative power
or not. For example this would permit bots that are either running as user or as
admin on Windows and run the test in both contexts.

Add a 'cygwin' dimension on Windows bots to be able to trigger on either inside
or outside a cygwin context.

Make get_attributes() a bit stronger in case of unexpected failure while
querying system details. This increases the chance for the bot to survive broken
start_slave.py.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/104990043