Add XsrfRemote for a simple automatically renewing XSRF protected client.
While this file is likely to be deleted soon as the bot code will be using
utils/net.py, I need it first before the transition. The mean reason I can't
convert to utils/net.py is because of POSTFORM, which is not a valid HTTP
request but "happens to work". I'll get rid of POSTFORM with the bot code
rewrite so the bot code rewrite effectively blocks the switch to utils/net.py.
Because of this, I cheaped out a bit on the url_helper unit test but I still
unit tested the class.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/106060043