Add pagination to /restricted/bot/bot_id.
Also change the default number of tasks from 10 to 100. It's more useful and
pretty much as fast as 10 items.

Print out more details like when the task ran and how much idle time the bot
wasted between tasks; we'd want a 100% busy bot.

The tasks where not properly sorted, making the table completely useless! Fix
this. Sadly, this require a new index.

Rework the UI to only show the Delete button on bots detected as missing. This
reduces the likelihood of clicking it on an healthy bot.

R=csharp@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/102540043