Add support for word-based task name search in /user/tasks.
This permits the user to search for a task via a keyword like 'Windows' or
'unit_tests'. The functionality uses the Search API. The indexing is done in a
way that do not interfere with the internal consistency of the DB even in case
of failure.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/101410044