Refactor ereporter2; split UI and backend API.
This will make extending the UI much simpler, which is much needed. I also plan
to add a generic error reporting facility to every services (in part for bug
119), so the code needs to be more organized to be reusable seamlessly in each
service.

Refactor path reformatting function as a standalone function in ereporter2 so it
can be reused eventually.

Make more code protected by prefixing with an underscore, which required moving
the code around a bit.

There is no functional change in this CL beside the standalone reformat_stack()
and adding './' and a elided path prefix.

R=vadimsh@chromium.org
BUG=swarming:119

Review URL: https://codereview.appspot.com/104390043