Change bot properties to use 'id' instead of 'tag'.
Change the id to be the base host name unqualified, e.g. keep 'foo' from host
'foo.bar.com'. The IDs do not need the FQDN.

Fix slave_machine.py to stop overriding 'id' consistently. The issue about the
network being up in time will be fixed later with a CL coming shortly, which
refactors this code significantly.

This change will rename all the bots with much shorter bot names, making it more
readable. So it is better to do this sooner than later. This also means
host.bar.com will conflict with host.foo.com. If this happens in practice, it is
up to the infrastructure people managing these hosts to change the id manually
in their start_slave.py.

The id can still be overriden, which is necessary for cases where multiple bots
are running on a single host, for example android testing.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/109420043