Change ereporter2 filtering to be more instance configurable.
Create an entity so that error categories can be temporarily or permanently
silenced, or that a minimum number of errors must be reported in the window for
the error to be reported. Using a threshold makes rare errors (like
InternalError) not trigger unless a lot of them occurs in the window, so that
'background noise' is not reported.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/106490044