Replace Results with TaskOutput to store tasks stdout.
The new structure is saved in the same root entity tree of TaskRequest so
cleaning up entities is easier.

The chunks are created in an monotonically increasing number and support
appending data so keys are predictable.

Reading the data doesn't require a query or stored keys, they are fetched
directly by key, which makes the DB operation faster.

This will be the basis to support stdout streaming.

R=csharp@google.com, csharp@chromium.org
BUG=swarming:116

Review URL: https://codereview.appspot.com/109000047