Logs the request parameters in case of Access Denied in Swarming.
This helps diagnose which astray bot that keeps on hammering the server. In
particular it is often the case that only the IPv6 address is not sufficient to
pinpoint the exact host running the unauthenticated bot. The POST parameters
usually includes the hostname, which helps a lot, to say the least.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/112920043