Fix an issue with partial writes in ndb.put_multi() with results update.
In this case, only one of TaskRunResult and TaskResultSummary is updated.
Sometimes the TaskOutputChunk is missing. In this case, the TaskResultSummary
may be in an invalid state. When the ndb.put_multi() call fails, the failure is
propagated up to the caller which will retry. But the code was seeing that the
TaskRunResult was already up to date and ignored the retry.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/112960044