Add android device properties in os_utilities.
This will later be leveraged to run Swarming android bots, if this is deemed a
good idea. So for now, this is just exploratory.

The general idea is that start_slave.py would decide if the bot should be an
android bot or a normal host bot, then act accordingly. The planned way to start
the Swarming bots is via udev rules (to be verified with udevadm), which is
relatively simple to write. The function get_adb_list_devices() would still be
useful on the initial bootstrap. lsusb -v will show the serial number as
iSerial which is the same as the adb devices reported device id.

There was still issues with this idea, namely when adb decides to hang
wholesale. The current approach doesn't address this issue, which may be a
blocker. We'd really want each device have its own adb daemon. One way would use
a chroot per Swarming bot - android device pair but I'm not knowledgeable enough
to know the feasibility.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/106700043