Fix a bug that was surfaced in 0a56b32346 that caused a crash loop.
The .duration property could be None when it shouldn't have. My hypothesis is
that a bot (likely a MBP) went to sleep long enough so that the
TaskRunResult.state was marked as BOT_DIED, then woke up back and completed the
task. This results in both .abandoned_ts and .completed_ts to be set.

R=csharp@chromium.org
BUG=

Review URL: https://codereview.appspot.com/118760044