Create /result2 to get rid of POSTFORM.
POSTFORM is not a valid http encoding. It just happened to work. Getting rid of
this is a prerequisite to switch over to the requests library. Once all bots are
updated, remove the old /result page.

Remove the need for common/swarm_constants.py.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/113380044