Add query to new client API.
This is part of the new client API. It will also be needed by the new bot API
unit test. The main advantage of this API is that it return results explicitly
either for the summary or for the run itself. This is explicitly to get rid of
/get_matching_test_cases, which relies on a query so it is flakier.

R=vadimsh@chromium.org
BUG=swarming:118

Review URL: https://codereview.appspot.com/120100043