Add new bot poll API.
Obviously this is not sufficient to run the bot but adding one API at a time
makes it easier to review and the unit test clearer.

After that, the task update API will be added. Once that's in place, the bot
code will be modified.

R=vadimsh@chromium.org
BUG=swarming:117

Review URL: https://codereview.appspot.com/116250043