Add 'quarantined' to sleep poll replies and add error endpoint.
The error endpoint permits easy quarantining from the bot itself, for example if
it fails to self-update. Refactor a bit the handler to be cleaner.

The new 'quarantine' flag simplify testing in the short term, as the bot will
likely not use this information for now. In the future it could be used to
disable functionalities like pre-warming but it's nowhere near for now.

R=csharp@chromium.org
BUG=swarming:117

Review URL: https://codereview.appspot.com/117590043