Add task_error handler for Bot-side error conditions after reaping a task.
This is to be used by slave_machine when local_test_runner fails, to ensure that
the task is properly marked as an infrastructure failure. Note that the cron job
would eventually figure it out, but if the bot knows and hasn't died yet, it's
much faster to have the bot actively notify the server right away.

Includes small refactoring to enable this functionality along new unit tests.

R=csharp@chromium.org
BUG=swarming:117

Review URL: https://codereview.appspot.com/116700043