Add links to the previous and following task on the same bot in task view.
This is helpful when debugging a task failure, in case a task left zombie
processes or other stateful changes behind.

Note that the link is to the TaskRunResult, not the TaskResultSummary, since
this is about tasks running on a bot, and a TaskResultSummary may refer to
multiple TaskRunResult, which may have run on separate bots.

Sadly, this requires more indexes than expected. I have ideas to reduce the
number of indexes but it'll require yet-another-refactor of the layout.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/133960043