Rewrite the bot code to use the new bot API.
The server has to be using the previous code for a while to ensure a smooth
transition, the idea is that the server supports both bot APIs simultaneously,
which makes the switch over and rollback seamless.

Convert the bot code from an OO design to a more functional one. It fits much
better the use case.

Call get_attributes() later. It was called too soon, potentially causing a race
condtiion when trying to get the FQDN while network is still configuring up.

R=csharp@chromium.org
BUG=swarming:117

Review URL: https://codereview.appspot.com/105240044