Add fake blinking cursor when a task is still streaming output.
This gives a clear visual indicator that stuff is going to happen when looking
at the bottom of a log. Right now it's a bit deceptive because the user has to
manually refresh the page to get an update, that will be fixed later.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/130690044