Fix dimensions_powerset_count(), it was miscalculating for strings.
This was made visible with the new bot code.

Clear up usage of Bot.quarantined. Make it clear that it is about a bot that
self-declared itself as quarantined. This saves a fetch on a mutating entity
type that was not needed. The proper fix is to use a parallel non-mutating
entity, this will scale much better with the DB.

Add logs to trap these cases in the logs.

Add bot_management import to remote_api.py

R=vadimsh@chromium.org
BUG=swarming:115
BUG=swarming:117

Review URL: https://codereview.appspot.com/132270043