Switch from a packet_number paradigm to chunk_start.
It simplifies the design and makes pretty much everything simpler. The code is
even shorter. The next step would be a full PUT REST API to send the stdout
independent of /task_update and a proper GET with range to support >32Mb
fetches.

This CL doesn't try to be a smooth upgrade path for the new bot API. This means
the canary will have its currently running tasks broken during the upgrade. This
is an non-issue for prod since it's currently using the old bot API.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/138020043