Move duplicate testing code to common file test_env_handlers.py.
Change the stock testing functions to use the new APIs instead of the old ones.
It's a requirement to get rid of the old APIs and to improve test coverage of
new features.

Remove the tests for the old client API. It's going away soon and removing the
tests right away helps the refactoring.

Stop uploading test_.+\.py on the server.

R=vadimsh@chromium.org, pgervais@chromium.org
BUG=swarming:118

Review URL: https://codereview.appspot.com/154450043