Add idempotency and task deduping support.
Do not assume idempotency, otherwise the heartbeat would never been executed
anymore! The client and recipes will be updated to specify it for tests.

Stop storing properties_hash in TaskRequest and add it to TaskResultSummary
instead. Only set it when the task is idempotent.

R=vadimsh@chromium.org, pgervais@chromium.org
BUG=swarming:163

Review URL: https://codereview.appspot.com/156280044