Store every version of bootstrap and bot_config instead of just the latest.
Add datastore_utils.append_decreasing() and store versioned for performance.

Replace file_chunks, which is removed. This is a breaking change, in the sense
that bot_config.py and bootstrap.py need to be reuploaded after updating a
version with this code but beforce switching it as the default.

Set bootstrap.py default encoding to utf-8.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/159170043