Fix bug in task deduplication; Add new swarming/api/v1/tasks client API.
When a task was deduped, TaskResultSummary was incorrectly copying the
.created_ts, .name and .user from the original TaskResultSummary.

The new API permits queries similar to /user/tasks. Refactor code from
handlers_frontend.py to task_result.py to be able to reuse more code.

A new ComputedProperty is necessary on TaskResultSummary to simplify the code,
which is still a TODO.

R=pgervais@chromium.org
BUG=

Review URL: https://codereview.appspot.com/159430043