Fix an issue with the error handler; issue with 5dbaaf9c.
get_remote() can be called multiple times during the course of a bot process
(unlike other tools) but it's guaranteed that the value is always the same, so
protect the call site accordingly.

This caught a new path in the code base, so update the test expectation.

Add bot.Bot.id to save from having similar problem in the future.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/164290044