Standardize bot event API; add BotEvent when bot is rebooting by itself.
Will give visibility on the time wasted in reboot, as it's the delta between the
event rebooting and the next connecting.

Merge all the messages of BotEvent to .message, it simplify things. We'll have
lost a few hours worth of events, it's not a big deal.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/172460043