Fix transaction in task_scheduler.py when expiring a task.
There is multiple issues in this file but I prefer to address each one at a
time. This is tricky code and easy to get wrong, so I prefer a lot of smaller
CLs. Sorry.

The end goal is that TaskResultSummary will always be consistent since always
fetched within a transaction.

Rename 'task' to 'to_run' when applicable to make it clear it's a reference to a
TaskToRun entity.

R=vadimsh@chromium.org
BUG=swarming:62

Review URL: https://codereview.appspot.com/175890043