Add more default bot dimensions; largest partition size and ram.
This is going to be used for perf testing, as I want to be able to able to run a
test on a specific kind of GCE instance, e.g. n1-standard-8 vs n1-highcpu-8.

Rename 'disk' to 'free_disks' to be clearer about its purpose.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/175120044