Fix task_update to use a transaction.
This should get rid of the 'Unexpected ordering' seen on the server likely due
to eventual consistency.

The last action to be converted to use a transaction is going to be
cancel_task(), as it's now going to be possible to cancel a running task. Ouf!

Include fix for pylint update.

R=vadimsh@chromium.org
BUG=swarming:62

Review URL: https://codereview.appspot.com/179940043