Wrap post_error() in try/except in run_bot() to catch extended network failure.
Looks like we lost a fair number of bots due to this error condition. The bot,
while running run_manifest() gets an exception, then it raises one while trying
to post the error, leading the process to crash. This try/except will keep the
loop alive so it continues polling even when the network is down for more than
12 minutes, e.g. twice the timeout per request, one for the original request,
another for the post_error() call.

Also improve resiliency when the network is down during the bot startup phase.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/180990043