Change the TaskRequest keys, creating a new primary task key.
This is changing an engine while flying. The trick used is the second to last
'0' in the task id, which is used to version the task ids.

The new format has an higher effective peak tasks/sec support of 64000
tasks/sec, effectively doubling randomness in the key from 8 bits to 16 bits,
which may be needed soon. The chromium infrastructure currently peaks at 5 TPS.
Reduce the project's life expectancy in 2014 from 4418 years to a mere 234
years. Key 4 bits for schema version control down from 8.

The most important benefit this change creates is that new tasks are appended
with smaller key id, meaning newest requests are naturally indexed first. This
will save a lot of indexes in index.yaml as there is no need to use created_ts
anymore. This will help with higher throughput by reducing the load on the DB's
index updater.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/178770043