Add user vs authenticated differentiation in TaskRequest.
The user is really just a free-form thing. authenticated is defined server-side
so it can't be spoofed. 'user' is still the most useful value. For example, we
want to know on behalf of who a bot sent a Swarming task for try job triggered
by the CQ on behalf of a user, we want to know which user it is. But
'authenticated' is important from an auditing standpoint.

For now, only display 'authenticated' on the task page.

Stop requiring user to be set. It's fine to be empty.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/187820043