Make default_expiration configurable via web UI. Change default to 30 days.
- Previous default was 7 days, now 30 days.
- Show last time the config was updated and by who.
- Add a test case for /restricted/config.

The cost of storage is very low compared to the I/O cost, so it's not going to
affect significantly the cost. On the other hand, keeping the files longer
solves problems when a user wants to see files from a task the week before or
with task deduplication.

R=vadimsh@chromium.org
BUG=chromium:445293

Review URL: https://codereview.appspot.com/191600043