Deny a bot retrying a task that it BOT_DIED in the inital try.
Everything was already there to do it, it was just not done yet. It'll greatly
help when a bot (usually Windows) gets into a killing spree and kills tasks in a
row, in the case where it would execute both tries #1 and #2, causing the task
to fail with BOT_DIED instead of simply being retried by another bot alive.

Updated test cases accordingly.

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/198260043