Do not refresh token immediately.
It was not necessary and it's an issue when bootstrapping via cygwin. What
happens is that third_party/oauth2client/multistore_file.py really wants to
write a token file right away and for some unknown reason the cygwin home
directory for some bots is not writable over ssh (I suspect an UAC issue).

R=vadimsh@chromium.org
BUG=

Review URL: https://codereview.appspot.com/197550043