Stop using test_env.py symlinks.
When PYTHONDONTWRITEBYTECODE=x is not defined, test_env.pyc is created aside a
test_env.py symlink. The problem is that the real test_env.py uses
os.path.realpath() to find its real path. On the second run, after a
test_env.pyc was created aside the test_env.py symlink, this one is loaded and
not the real source file. This means that the source file path is different,
causing generalized failure.

(CL was reviewed manually, Rietveld couldn't handle replacing a symlink with a
file)

R=vadimsh@chromium.org
BUG=