utils.memcache: key arg whitelist
Added key_args parameter which is a whitelist of name of args to include in the
cache key. This is needed for cases when a value of one arg ultimately
depends on another arg, and its repr is not deterministic. For example, if you pass a function argument.

It also simplifies how
self and cls args are handled.

R=maruel@chromium.org, vadimsh@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1069873002