Add support for using service accounts with private keys.
Sometimes GAE's app_identity.get_access_token(...) is not enough. For example,
if GAE app needs to use GCE API, it needs to be added to corresponding Cloud
project, which is not always possible. On the other hand, a service account
created for a project is added to the project by default.

Another feature of this CL is support for using "fake" tokens from datastore on
dev server. It allows to debug apps that make API calls on dev server using
developer's own OAuth2 token (after it is added to the local datastore).

Finally, change 'OAuth %s' authorization header to more standard 'Bearer %s'.

R=maruel@chromium.org, nodir@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1077323002