Handle TombstonedTaskError in utils.enqueue_task.
TombstonedTaskError is raised when task names are used to deduplicate tasks. It
indicated that the task was not added because a task with such name was added
already. In all current application (in infra.git) it is not a error, but
current produced WARNING-level log message. This CL fixes that.

Also add 'countdown' parameter, useful for scheduling tasks to run later.

R=nodir@chromium.org
TEST=deployed, works

Review URL: https://codereview.chromium.org/1152073003