Separate machines and capacity in the Machine Provider's catalog
Machine entries are updated with add_machine and delete_machine. Hostname is required, and uniqueness is enforced per-backend. The backend+hostname combination is used as the ID.

Capacity entries are updated with a single modify_capacity call. The backend+dimensions combination is used as the ID, but there is also a count parameter which indicates how many machines with the given dimensions the backend has capacity for. Technically, there should be no hostname given here, but this is not enforced. Presumably the backend could be pre-allocating hostnames or whatever.

Backends can "delete" capacity by setting count to 0. This doesn't remove it from the datastore though (maybe it should; can be done later on).

BUG=489837
TESTED=handlers_endpoints_test.py and https://1464-597b3d1-tainted-smut-dot-machine-provider.appspot.com/_ah/api/explorer
R=maruel@chromium.org

Review URL: https://codereview.chromium.org/1216613009 .